roles/docker_fastapi/files/app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
prometheus-fastapi-instrumentator==6.1.0